fixing bug with pkgconf in ubuntu 16.04
For some reason, an update to spack resulted in pkgconf being needed
and failing because it uses format strings (not available in Python
3.5 in the image). I think this might be because this version
uses clingo as the default solver, and bootstraps it, and it possibly
worked before because we just did not do that. The hacky solution is
to install a newer version of python in the image (3.8) and try again.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>