Update hip copy path
Given the changes for the build to be relative to the root, this copy needs to be updated. I think we are using {prefix} that has -f Dockerfile but if we define it twice the second one might take. If needed we could tweak this in uptodate so it's not so buggy.